fix: wrap useSearchParams in Suspense boundary for Next.js 15 compatibility
